ml/requirements.txt
# ML dependencies for volatility prediction
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
psycopg2-binary>=2.9.0
